# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/music/artist/{name}": {
            "get": {
                "description": "Retrieve music information by the name of the artist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "music"
                ],
                "summary": "Get music by artist name",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"John Doe\"",
                        "description": "Name of the artist",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved music information",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "400": {
                        "description": "Name cannot be empty",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Music not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/music/{isrc}": {
            "get": {
                "description": "Retrieve music information by its ISRC (International Standard Recording Code)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "music"
                ],
                "summary": "Get music by ISRC",
                "parameters": [
                    {
                        "type": "string",
                        "format": "isrc",
                        "description": "ISRC of the music",
                        "name": "isrc",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved music information",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "400": {
                        "description": "ISRC not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Music not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tracks/{isrc}": {
            "post": {
                "description": "Retrieve track information from Spotify based on the provided ISRC",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Insert music track metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ISRC of the track",
                        "name": "isrc",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Track inserted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "ISRC not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Track not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}